# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Auto Ecole documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/agence": {
            "get": {
                "tags": [
                    "agences"
                ],
                "summary": "Get a list of agences",
                "operationId": "09cc2e0c34631bf0a618e6893c634f5a",
                "responses": {
                    "200": {
                        "description": "List of agences"
                    }
                }
            },
            "post": {
                "tags": [
                    "agences"
                ],
                "summary": "POST to create agences",
                "operationId": "4c40206458786b9a9a2d0a03d51aa4d1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "fullname": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "email"
                                    },
                                    "password": {
                                        "type": "password"
                                    },
                                    "adresse": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "inmmatriculation": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Create agences",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/agence/{id}": {
            "put": {
                "tags": [
                    "agences"
                ],
                "summary": "Update an agence",
                "operationId": "8f6382a2fa6fc7af111fb38a12dcabf6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the agence to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "fullname": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "adresse": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "inmmatriculation": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Agence updated successfully"
                    },
                    "404": {
                        "description": "Agence not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "agences"
                ],
                "summary": "Delete an agence",
                "operationId": "5220bb774319d9dc7ea1200b4087f00e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the agence to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Agence deleted successfully"
                    },
                    "404": {
                        "description": "Agence not found"
                    }
                }
            }
        },
        "/api/clients": {
            "get": {
                "tags": [
                    "Client"
                ],
                "summary": "Get a list of clients",
                "operationId": "0b636ca85e574ff3010b44d6c621c537",
                "responses": {
                    "200": {
                        "description": "List of clients"
                    }
                }
            },
            "post": {
                "tags": [
                    "Client"
                ],
                "summary": "POST to create clients",
                "operationId": "9fdaf88d446f92c97501f370f36678cb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "fullname": {
                                        "type": "string"
                                    },
                                    "conatct": {
                                        "type": "string"
                                    },
                                    "adresse": {
                                        "type": "string"
                                    },
                                    "age": {
                                        "type": "integer"
                                    },
                                    "agence_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Create clients",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/clients/{id}": {
            "put": {
                "tags": [
                    "Client"
                ],
                "summary": "Update an clients",
                "operationId": "9ffd4acb0368eb6068e75cd7dd5d43ff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the clients to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "fullname": {
                                        "type": "string"
                                    },
                                    "conatct": {
                                        "type": "integer"
                                    },
                                    "adresse": {
                                        "type": "string"
                                    },
                                    "age": {
                                        "type": "integer"
                                    },
                                    "agence_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Client updated successfully"
                    },
                    "404": {
                        "description": "Client not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Client"
                ],
                "summary": "Delete an agence",
                "operationId": "527e3694096ac0e9d71511de8b647c8d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the clients to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "clients deleted successfully"
                    },
                    "404": {
                        "description": "clients not found"
                    }
                }
            }
        },
...
    }
}